Initial support for replication configuration changes in PostgreSQL 12
As there's no longer a replication configuration file in a known location
(i.e. $PGDATA/recovery.conf), for this initial implementation we'll require
use of a dedicated configuration file (the location of which can be
optionally specified) to store the replication settings generated by
repmgr.

This does however rely on the user ensuring this file is included from
the main PostgreSQL configuration.

In future iterations we'll try and make this more robust.